hk integration

+++ apps/adcsttMC_app/fsw/inc/adcsttMC_app_msg.h
@@ -46,10 +46,10 @@ typedef struct
     uint8              CmdCounter;          /**< \brief Count of valid commands received */
     uint8              ErrCounter;          /**< \brief Count of invalid commands received */
     uint8              Reserved[2];        /**< \brief Reserved bytes to fill compiler padding */
-    double             quaternion1;         /**< \brief Quaternion component 1 */
-    double             quaternion2;         /**< \brief Quaternion component 2 */
-    double             quaternion3;         /**< \brief Quaternion component 3 */
-    double             quaternion4;         /**< \brief Quaternion component 4 */
+    float              quaternion1;         /**< \brief Quaternion component 1 */
+    float              quaternion2;         /**< \brief Quaternion component 2 */
+    float              quaternion3;         /**< \brief Quaternion component 3 */
+    float              quaternion4;         /**< \brief Quaternion component 4 */
 } ADCSTTMC_HkTlm_Adcstt_t;
 
 /**
